requirements.txt
# Core dependencies for AgriToken backend
algokit-utils==4.2.0
python-dotenv==1.1.1
algorand-python==2.9.0
algorand-python-testing==0.5.0

# Flask API dependencies
Flask==3.0.0
Flask-CORS==4.0.0

# Dependencies for algokit-utils
httpx==0.28.1
py-algorand-sdk==2.10.0
typing-extensions==4.15.0

# Dependencies for httpx
anyio==4.10.0
certifi==2025.8.3
httpcore==1.0.9
idna==3.10
sniffio==1.3.1

# Dependencies for py-algorand-sdk
msgpack==1.1.1
pycryptodomex==3.23.0
pynacl==1.5.0

# Dependencies for algorand-python-testing
coincurve==21.0.0
ecdsa==0.19.1
six==1.17.0

# Dependencies for coincurve
cffi==1.17.1

# Dependencies for cffi
pycparser==2.22

# Dependencies for pynacl
cffi==1.17.1

# Dependencies for ecdsa
six==1.17.0
